fix: correct engine rectangle usage
The current code creates a new rectangle rather than using the engine's
rectangle, adds references to its buffer to the deletion queue, then
deletes the rectangle when the procedure returns.

First, this causes a compiler error. The deletion queue can handle an
Allocated_Buffer, but not a pointer to one. We should queue the buffers
by value, instead.

Next, if we don't set the engine's rectangle property, we'll get a hang
and then a crash with a few Validation Layer errors. This is because
we're trying to draw an uninitialized index. Using `self.rectangle`
instead of `rectangle` solves this issue.